pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "velodoc-llm-train"
version = "0.1.0"
description = "Velodoc LLM training pipeline (Hydra + MLflow + HF Hub)"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "rich>=13.7.0",
  "tqdm>=4.66.0",
  "numpy>=1.26.0",
  "pydantic>=2.6.0",
  "datasets>=2.20.0",
  "transformers>=4.44.0",
  "accelerate>=0.33.0",
  "peft>=0.12.0",
  "trl>=0.9.6",
  "mlflow>=2.14.0",
  "huggingface_hub>=0.24.0",
  "safetensors>=0.4.3",
]

[project.optional-dependencies]
deepspeed = ["deepspeed>=0.14.4"]

[project.scripts]
velodoc-train = "velodoc_train.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
